Closes #34 again: update datastructure test for float64 datatype

+++ tests/test_datastructure.py
@@ -120,14 +120,14 @@ class TestDatastructure(unittest.TestCase):
         self.assertEqual(len(w.samples),
                          len(samples))
         self.assertIsInstance(w.samples, np.ndarray)
-        self.assertEqual(w.samples.dtype, np.float32)
+        self.assertEqual(w.samples.dtype, np.float64)
 
-        w.samples = np.array(samples, dtype=np.float32)
+        w.samples = np.array(samples, dtype=np.float64)
 
         self.assertEqual(len(w.samples),
                          len(samples))
         self.assertIsInstance(w.samples, np.ndarray)
-        self.assertEqual(w.samples.dtype, np.float32)
+        self.assertEqual(w.samples.dtype, np.float64)
 
     def tearDown(self):
         pass
